added some logic

+++ sh/nvim.sh
@@ -1,17 +1,29 @@
-sudo apt-get install fuse libfuse2 git python3-pip ack-grep -y
-wget --quiet https://github.com/neovim/neovim/releases/download/nightly/nvim.appimage --output-document nvim
-chmod +x nvim
-sudo chown root:root nvim
-sudo mv nvim /usr/bin
-cd ~
-mkdir -p .config/nvim
-curl -fLo ~/.local/share/nvim/site/autoload/plug.vim --create-dirs https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
-pip3 install --user neovim
-touch ~/.config/nvim/init.vim
-echo -e"call plug#begin()\n\
+#Install Nvim if needed and update plugin configuration
+plugs="call plug#begin()\n\
 Plug 'roxma/nvim-completion-manager'\n\
 Plug 'SirVer/ultisnips'\n\
 Plug 'honza/vim-snippets'\n\
-call plug#end()"
-nvim
+Plug 'jacobsimpson/nvim-example-python-plugin'\n\
+Plug 'stephpy/vim-yaml'\n\
+call plug#end()" 
+
+
+if ! [ -f $(which nvim) ]
+then
+	sudo apt-get install fuse libfuse2 git python3-pip ack-grep -y
+	wget --quiet https://github.com/neovim/neovim/releases/download/nightly/nvim.appimage --output-document nvim
+	chmod +x nvim
+	sudo chown root:root nvim
+	sudo mv nvim /usr/bin
+	pip3 install --user neovim
+	curl -fLo ~/.local/share/nvim/site/autoload/plug.vim --create-dirs https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
+fi
+
+cd ~
+mkdir -p .config/nvim
+touch ~/.config/nvim/init.vim
+echo -e $plugs > ~/.config/nvim/init.vim
+
+sed -i.bu 's/async=True/**{"async": True}/' ~/.config/nvim/plugged/nvim-completion-manager/pythonx/cm.py
+sed -i.bu 's/async=True/**{"async": True}/' ~/.config/nvim/plugged/nvim-completion-manager/pythonx/cm_core.py
